+++ src/components/Playlist/PlaylistHeader/SongSearchbar.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import TextField from '@mui/material/TextField';
 
 interface Props {
